Remove `GraphqlClient` trait (#58)
I've been working with `graphql-ws-client` recently and I'm not that
happy with it's API. This is the first in a series of refactors to
simplify it a bunch.

First on the chopping block: the `GraphqlClient` trait. There wasn't a
ton of point to this - it lets client implementations avoid temporarily
deserializing _errors_ into `serde_json::Value` but generally the
response data still has to go through this. I suppose it'd also let
clients use something other than `serde_json` but in practice I'm not
sure how useful this is.

Removing this simplifies the API a bunch, so I think it's worthwhile.