Implement the new API (#61)
As outlined in #59 - I'm not happy with the current API.

This PR introduces most of the new API I want to support - this is
basically ended up as a rewrite rather than the refactor I'd intended.
But that has some advantages - I'm probably going to release v0.8.0 with
the old API still intact but marked as deprecated. I think this might
provide a nicer update experience than being faced with an immediate
wall of compiler errors.

Future PRs will continue with the rest of the work.