An easier one shot subscribe API (#65)
`graphql-ws-client` was designed to allow multiple subscriptions to be
run on a single connection. This is still something I want to support,
but it's also common for users to only want to run a single subscription
per connection. Right now the API for doing that is a bit too verbose.

This PR will introduce a `streaming_operation` function on the
`ClientBuilder` to handle this use case, combining the connect,
subscribe and actor spawning into a single function call.

Todo:
- [x] Examples
- [x] Tests
- [x] Changelog