chore: add release-plz config
I wanted to use release-plz without config in this repo, but it seems like it
doesn't pay attention to `publish = "false"` when determining whether or not
there's a single public crate in the workspace.

This was leading to it not using my preferred scheme for release tags.  I'd
maybe like to fix upstream but for now it is far easier to just configure it.