feat!: Connection functions now return impl Future
Now that the MSRV is > 1.75 we can start making use of return position impl trait.

We need to specify send bounds on the returned futures so we return an impl
Trait, but it seems like rust is smart enough to allow us to use an actual
async function in the implementations.