fix: remove futures from ws_stream_wasm module
In #107 I removed futures in favour of futures-lite in various places.
However, because ws_stream_wasm is not a default feature I didn't notice that
I'd not removed futures from there as well.

CI also did not catch this because futures was used for the tungstenite
feature, which is often always on.

This PR removes futures entirely, by implementing our own copy of
StreamExt::send based on the futures impl, and then making sure there's no
remaining uses of futures in the crate.