fix: infinite loop (#124)
Fixes #123 

There might be an edge case I am not thinking about, but my reasoning is
that if all clients have disconnected then there is no point in keeping
the operations alive. Also we keep a clone of the client in all
Subscription objects so this means the channel wont be closed until all
streams are dropped.