feat: add `Client.stop(subscription_id)`
Maintaining a list of subscriptions on a connection to be able to stop
them with a list of Subscription items is difficult/complex due to the
generics. I couldn't figure out how to box them when I am sending
different subscription types.

Instead just exposing the subscription id and allowing you to cancel the
subscription by id on the client itself allows you to manage
subscriptions easily and not have to worry about the request types at
all.